Phase 4.3: Complete forms.c and move.c -Wall warning elimination
Prompt: Complete Phase 4.3 warning elimination for forms.c and move.c files

Work Completed:
✅ forms.c (Game-only file):
- Fixed function declaration conflicts: Removed conflicting fopen() declarations (2 locations)
- Fixed format warnings: Changed BREAKJIHAD from %d to %ld format (3 locations)
- Fixed format warning: Changed TURN from %ld to %d format (1 location)
- Fixed buffer safety: sprintf overlap warning with temporary variable (line 988)
- Fixed buffer safety: sprintf overflow warning with snprintf conversion (line 1371)
- Result: Zero -Wall warnings achieved

✅ move.c (Game-only file):
- Fixed dangling else warnings: Added explicit braces (2 locations: lines 309, 624)
- Fixed format warnings: Changed groupmen and TAKESECTOR from %d to %ld (line 645)
- Result: Zero -Wall warnings achieved

📊 Progress Update:
- Phase 4.3: 11/34 files complete (32.4% progress)
- Files completed: admin.c, check.c, psmap.c, spew.c, combat.c, newlogin.c, update.c, makeworl.c, main.c, forms.c, move.c
- Remaining: 23 files with warnings to eliminate

🔧 Technical Patterns Applied:
- Function declaration conflicts: Remove redundant local declarations
- Format safety: Match format specifiers to argument types (%ld for long, %d for int/short)
- Dangling else: Add explicit braces for clarity
- Buffer safety: sprintf → snprintf with proper sizing
- Compilation validation: Game-only flags (-DCONQUER) for both files

🧪 All files tested and verified to compile with zero -Wall warnings

+++ psmap.c
@@ -927,7 +927,7 @@ main(int argc, char *argv[])
 	fprintf(stderr, "Not a Conquer Map file\n");
 	exit(1);
     }
-    for (c = 0; c < strlen(firstline); c++) {
+    for (c = 0; c < (int)strlen(firstline); c++) {
 	if (firstline[c] == ':')
 	    break;
     }
